Create Assignment_A_(07)
Write a PL/SQL block of code using parameterized Cursor that will merge the data available in
the newly created table N_Roll Call with the data available in the table O_RollCall. If the data
in the first table already exist in the second table then that data should be skipped.